Tests/LibWeb/Text/expected/css/font-face-width-matching.txt

reference fonts differ: true
condensed uses HashSans: true
normal uses Lato: true
